fix(retime): unsigned audio tempo and fail-closed timeMapRange
Rust exposes resolveTimeMapAudioTempo, the unsigned magnitude of the
effective rate, so a reverse ramp never reaches SoundTouch as a negative
tempo; the web adapter reads it through a small tested helper. Rust trim
now rejects timeMapRange on a rate-based clip instead of silently ignoring
it. Transport tick fields are integers throughout the time-map schema, and
a windowed-audio regression test covers slicing a reverse map to a nonzero
preview window with source trims held fixed.

Refs #26

+++ packages/mcp-server/src/tool-schemas.ts
@@ -5,11 +5,12 @@ import { QC_CHECK_IDS } from "./export-qc";
 import { BOOTSTRAP_PROJECT_ID } from "./managed-editor-worker";
 import { operationIdSchema } from "./operation-tool-schemas";
 
-// This schema owns the JSON transport shape only: the wire-format version tag
-// and field layout that MCP clients see in the published tool schema. Rust
-// validates all canonical time-map semantics, including tick integrality/range,
-// rates, fallback policy, continuity, and non-empty identities/segments.
-const timeMapTickSchema = z.number();
+// This schema owns the JSON transport shape only: the wire-format version tag,
+// integer tick fields, and the field layout that MCP clients see in the
+// published tool schema. Rust validates all canonical time-map semantics,
+// including tick ranges, rates, fallback policy, continuity, and non-empty
+// identities/segments.
+const timeMapTickSchema = z.number().int();
 const timeMapRateSchema = z.number();
 
 const timeMapSpeedSegmentSchema = z.object({
